mpipe_requirements.txt
absl-py==0.12.0
attrs==20.3.0
dataclasses==0.6
imutils==0.5.4
numpy==1.20.2
opencv-contrib-python==4.5.1.48
# opencv-python==4.5.1.48
mediapipe==0.8.3.1
protobuf==3.15.7
six==1.15.0
tqdm==4.59.0
pyyaml